# openapi/personalcrm_openapi_v1.yaml
openapi: 3.0.3
info:
  title: PersonalCRM API
  version: 1.0.0
  description: |
    PersonalCRM v1 API for managing contacts, interactions, introductions, reminders,
    strategy/wheel, projects, and RBAC in a cloud workspace with assistant access.

    Notes:
    - All requests are scoped to a workspace.
    - Field-level privacy is enforced server-side; assistants may receive redacted fields.
    - Use cursor-based pagination (next_cursor).
servers:
  - url: https://api.personalcrm.example.com
    description: Production (replace)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/workspaces:
    get:
      tags: [RBAC]
      summary: List workspaces available to the user
      operationId: listWorkspaces
      responses:
        '200':
          description: Workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workspaces/{workspace_id}/members:
    get:
      tags: [RBAC]
      summary: List workspace members
      operationId: listWorkspaceMembers
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkspaceMember'
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [RBAC]
      summary: Invite/add a workspace member
      operationId: createWorkspaceMember
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceMemberCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/workspaces/{workspace_id}/members/{user_id}:
    patch:
      tags: [RBAC]
      summary: Update member role/status
      operationId: updateWorkspaceMember
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceMemberUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/reminders:
    get:
      tags: [Reminders]
      summary: List reminders
      operationId: listReminders
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReminderStatus'
        - name: due_from
          in: query
          schema:
            type: string
            format: date-time
        - name: due_to
          in: query
          schema:
            type: string
            format: date-time
        - name: contact_id
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Reminders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReminders'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Reminders]
      summary: Create a reminder
      operationId: createReminder
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/reminders/{reminder_id}:
    patch:
      tags: [Reminders]
      summary: Update a reminder
      operationId: updateReminder
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ReminderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/strategy:
    get:
      tags: [Strategy]
      summary: Get workspace strategy
      operationId: getStrategy
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '200':
          description: Strategy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Strategy]
      summary: Upsert workspace strategy
      operationId: upsertStrategy
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyUpsert'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/strategy/wheel:
    get:
      tags: [Strategy]
      summary: Get wheel-of-environment segments
      operationId: getStrategyWheel
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '200':
          description: Wheel
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WheelSegment'
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Strategy]
      summary: Replace all wheel segments
      operationId: putStrategyWheel
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/WheelSegmentUpsert'
              required: [data]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WheelSegment'
                required: [data]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceIdQuery:
      name: workspace_id
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
      description: Workspace scope for the request
    WorkspaceIdPath:
      name: workspace_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
    UserIdPath:
      name: user_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    ReminderIdPath:
      name: reminder_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    UUID:
      type: string
      format: uuid
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required: [code, message]
      required: [error]
...
    ReminderType:
      type: string
      enum: [follow_up, birthday, anniversary, task, custom]
    ReminderStatus:
      type: string
      enum: [open, done, canceled]
...
    Workspace:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, name, created_at]
    WorkspaceMember:
      type: object
      properties:
        workspace_id: { $ref: '#/components/schemas/UUID' }
        user_id: { $ref: '#/components/schemas/UUID' }
        email: { type: string, format: email }
        display_name: { type: string }
        membership_role:
          type: string
          enum: [owner, assistant, collaborator]
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [workspace_id, user_id, membership_role, is_active, created_at]
    WorkspaceMemberCreate:
      type: object
      properties:
        email: { type: string, format: email }
        membership_role:
          type: string
          enum: [assistant, collaborator]
      required: [email, membership_role]
    WorkspaceMemberUpdate:
      type: object
      properties:
        membership_role:
          type: string
          enum: [owner, assistant, collaborator]
        is_active: { type: boolean }
...
    Reminder:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        contact_id: { $ref: '#/components/schemas/UUID', nullable: true }
        project_id: { $ref: '#/components/schemas/UUID', nullable: true }
        type: { $ref: '#/components/schemas/ReminderType' }
        status: { $ref: '#/components/schemas/ReminderStatus' }
        title: { type: string }
        body: { type: string }
        due_at: { type: string, format: date-time }
        assigned_to_user_id: { $ref: '#/components/schemas/UUID', nullable: true }
        created_by_user_id: { $ref: '#/components/schemas/UUID' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, type, status, due_at, created_at, updated_at]
    ReminderCreate:
      type: object
      properties:
        contact_id: { $ref: '#/components/schemas/UUID' }
        project_id: { $ref: '#/components/schemas/UUID' }
        type: { $ref: '#/components/schemas/ReminderType' }
        title: { type: string }
        body: { type: string }
        due_at: { type: string, format: date-time }
        assigned_to_user_id: { $ref: '#/components/schemas/UUID' }
      required: [type, due_at]
    ReminderUpdate:
      type: object
      properties:
        status: { $ref: '#/components/schemas/ReminderStatus' }
        title: { type: string }
        body: { type: string }
        due_at: { type: string, format: date-time }
        assigned_to_user_id: { $ref: '#/components/schemas/UUID', nullable: true }
    Strategy:
      type: object
      properties:
        workspace_id: { $ref: '#/components/schemas/UUID' }
        vision: { type: string }
        goals:
          type: array
          items:
            type: object
            additionalProperties: true
        swot:
          type: object
          additionalProperties: true
        roadmap:
          type: object
          additionalProperties: true
        updated_at: { type: string, format: date-time }
      required: [workspace_id, updated_at]
    StrategyUpsert:
      type: object
      properties:
        vision: { type: string }
        goals:
          type: array
          items:
            type: object
            additionalProperties: true
        swot:
          type: object
          additionalProperties: true
        roadmap:
          type: object
          additionalProperties: true
    WheelSegment:
      type: object
      properties:
        workspace_id: { $ref: '#/components/schemas/UUID' }
        key:
          type: string
          description: Segment identifier (e.g., health, family, career...)
        current:
          type: integer
          minimum: 0
          maximum: 10
        target:
          type: integer
          minimum: 0
          maximum: 10
        notes: { type: string }
        updated_at: { type: string, format: date-time }
      required: [workspace_id, key, current, target, updated_at]
    WheelSegmentUpsert:
      type: object
      properties:
        key: { type: string }
        current:
          type: integer
          minimum: 0
          maximum: 10
        target:
          type: integer
          minimum: 0
          maximum: 10
        notes: { type: string }
      required: [key, current, target]
...
    PaginatedReminders:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Reminder' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...